requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
flake8==6.0.0
idna==3.4
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
telegram==0.0.1
tornado==6.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
